@@ -2,7 +2,7 @@
 name = "libprettylogger"
 version = "3.0.2"
 authors = ["tpaau-17DB"]
-edition = "2021"
+edition = "2024"
 description = "Fancy logger library."
 readme = "README.md"
 homepage = "https://github.com/tpaau-17DB/libprettylogger"
@@ -12,7 +12,6 @@ keywords = ["logger", "customizable", "log-levels", "log-filtering", "log-format
 categories = ["development-tools"]
 
 [dependencies]
-lazy_static = "^1.0"
 chrono = "^0.4"
 glob = "^0.3"
 serde = { version = "^1.0", features = ["derive"] }
