@@ -5,8 +5,52 @@ use std::sync::{
 
 use crate::Logger;
 
-/// Global `Logger` struct that can be used with the `debg!`, `info!`, `warn!`,
+/// Global `Logger` struct that can be used with the `debug!`, `info!`, `warn!`,
 /// `err!`, and `fatal!` macros.
+///
+/// > **Note**
+/// > Do not use any log macros when you are writing to the global logger, it
+/// > will cause your thread to lock.
+///
+/// This will block the thread:
+/// ```no_run
+/// # use prettylogger::{info, glob::LOGGER};
+/// // Get write access to the logger
+/// let mut logger = LOGGER.write().unwrap();
+///
+/// // Trying to use logging macro blocks the thread
+/// info!("This will never be shown!");
+/// ````
+///
+/// # Examples
+///
+/// Using global logging:
+/// ```
+/// # use prettylogger::{debug, info, warn, err, fatal};
+/// debug!("This is a debug message!");
+/// info!("This is an info message!");
+/// warn!("This is a warning!");
+/// err!("This is an error!");
+/// fatal!("This is a fatal error!");
+/// ```
+///
+/// Configuring the global logger:
+/// ```
+/// # use prettylogger::{
+/// #     config::Verbosity,
+/// #     glob::LOGGER,
+/// #     debug
+/// # };
+/// // Configure logger
+/// let mut logger = LOGGER.write().unwrap();
+/// logger.set_verbosity(Verbosity::All);
+///
+/// // Drop the logger so a read lock can be acquired
+/// drop(logger);
+///
+/// // Then print a message
+/// debug!("This should be shown!");
+/// ```
 pub static LOGGER: LazyLock<RwLock<Logger>>
     = LazyLock::new(|| RwLock::new(Logger::default()));
 
@@ -16,20 +60,22 @@ pub static LOGGER: LazyLock<RwLock<Logger>>
 /// Not to be confused with the `dbg!` macro.
 ///
 /// > **Warning!**
-/// > This macro will block if any thread holds read access to the global
-/// > logger’s RW lock.
+/// > This macro will block if any thread holds write access to the global
+/// > logger.
 ///
 /// # Panics
 /// Panics if the global logger's lock is poisoned.
 ///
 /// # Examples
-/// ```rust
-/// use prettylogger::debg;
+///
+/// Using the `debug!` macro:
+/// ```
+/// use prettylogger::debug;
 /// let name = String::from("world");
-/// debg!("Hello, {name}!");
+/// debug!("Hello, {name}!");
 /// ```
 #[macro_export]
-macro_rules! debg {
+macro_rules! debug {
     ($($t:tt)*) => {{
         use $crate::glob::LOGGER;
         LOGGER
@@ -42,14 +88,16 @@ macro_rules! debg {
 /// Prints an info message using the global `Logger` instance.
 ///
 /// > **Warning!**
-/// > This macro will block if any thread holds read access to the global
-/// > logger’s RW lock.
+/// > This macro will block if any thread holds write access to the global
+/// > logger.
 ///
 /// # Panics
 /// Panics if the global logger's lock is poisoned.
 ///
 /// # Examples
-/// ```rust
+///
+/// Using the `info!` macro:
+/// ```
 /// use prettylogger::info;
 /// let name = String::from("world");
 /// info!("Hello, {name}!");
@@ -65,17 +113,19 @@ macro_rules! info {
     }};
 }
 
-/// Prints a warning message using the global `Logger` instance.
+/// Prints a warning using the global `Logger` instance.
 ///
 /// > **Warning!**
-/// > This macro will block if any thread holds read access to the global
-/// > logger’s RW lock.
+/// > This macro will block if any thread holds write access to the global
+/// > logger.
 ///
 /// # Panics
 /// Panics if the global logger's lock is poisoned.
 ///
 /// # Examples
-/// ```rust
+///
+/// Using the `warn!` macro:
+/// ```
 /// use prettylogger::warn;
 /// let name = String::from("world");
 /// warn!("Hello, {name}!");
@@ -91,17 +141,19 @@ macro_rules! warn {
     }};
 }
 
-/// Prints an error message using the global `Logger` instance.
+/// Prints an error using the global `Logger` instance.
 ///
 /// > **Warning!**
-/// > This macro will block if any thread holds read access to the global
-/// > logger’s RW lock.
+/// > This macro will block if any thread holds write access to the global
+/// > logger.
 ///
 /// # Panics
 /// Panics if the global logger's lock is poisoned.
 ///
 /// # Examples
-/// ```rust
+///
+/// Using the `err!` macro:
+/// ```
 /// use prettylogger::err;
 /// let name = String::from("world");
 /// err!("Hello, {name}!");
@@ -117,17 +169,19 @@ macro_rules! err {
     }};
 }
 
-/// Prints a fatal error message using the global `Logger` instance.
+/// Prints a fatal error using the global `Logger` instance.
 ///
 /// > **Warning!**
-/// > This macro will block if any thread holds read access to the global
-/// > logger’s RW lock.
+/// > This macro will block if any thread holds write access to the global
+/// > logger.
 ///
 /// # Panics
 /// Panics if the global logger's lock is poisoned.
 ///
 /// # Examples
-/// ```rust
+///
+/// Using the `fatal!` macro:
+/// ```
 /// use prettylogger::fatal;
 /// let name = String::from("world");
 /// fatal!("Hello, {name}!");
